run/requirements.txt
docker-compose
pika
six
